Moss-live-labs/examples/image-search/backend-py/pyproject.toml
[project]
name = "moss-image-search-demo"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "inferedge-moss==1.0.0b19",
    "python-dotenv>=1.2.1",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "boto3>=1.34.0",
]
